# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/register": {
      "post": {
        "description": "Creates a new user account and returns user info with JWT access token.",
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 201,
                    "message": "Operation successfully!",
                    "data": {
                      "user": {
                        "id": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
                        "email": "jane.doe@example.com",
                        "name": "Jane Doe"
                      },
                      "accessToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Register a new user",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "description": "Authenticates with email and password, returns user info and JWT access token.",
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 201,
                    "message": "Operation successfully!",
                    "data": {
                      "user": {
                        "id": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
                        "email": "jane.doe@example.com",
                        "name": "Jane Doe"
                      },
                      "accessToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Login",
        "tags": [
          "auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "Starter Kit API Documentation",
    "description": "API documentation for Starter Kit",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "jwt": {
        "scheme": "Bearer",
        "bearerFormat": "Bearer",
        "description": "[just text field] Please enter token in following format: Bearer <JWT>",
        "name": "Authorization",
        "type": "http",
        "in": "Header"
      }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "The email of the user",
            "example": "jane.doe@example.com"
          },
          "password": {
            "type": "string",
            "description": "The password of the user",
            "example": "password"
          },
          "name": {
            "type": "string",
            "description": "The name of the user",
            "example": "Jane Doe"
          }
        },
        "required": [
          "email",
          "password",
          "name"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "The email of the user",
            "example": "jane.doe@example.com"
          },
          "password": {
            "type": "string",
            "description": "The password of the user",
            "example": "password"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
...
    }
  },
  "security": [
    {
      "jwt": []
    }
  ]
}